+++ src/Firebase/Messaging/ApnsConfig.php
@@ -8,9 +8,13 @@ use JsonSerializable;
 
 /**
  * @see https://developer.apple.com/documentation/usernotifications/setting_up_a_remote_notification_server/generating_a_remote_notification
+ * @see https://developer.apple.com/documentation/usernotifications/setting_up_a_remote_notification_server/sending_notification_requests_to_apns
  */
 final class ApnsConfig implements JsonSerializable
 {
+    private const PRIORITY_CONSERVE_POWER = '5';
+    private const PRIORITY_IMMEDIATE = '10';
+
     /** @var array<string, mixed> */
     private $config;
 
@@ -63,6 +67,25 @@ final class ApnsConfig implements JsonSerializable
         $config->config['payload'] = $config->config['payload'] ?? [];
         $config->config['payload']['aps'] = $config->config['payload']['aps'] ?? [];
         $config->config['payload']['aps']['badge'] = $number;
+
+        return $config;
+    }
+
+    public function withImmediatePriority(): self
+    {
+        return $this->withPriority(self::PRIORITY_IMMEDIATE);
+    }
+
+    public function withPowerConservingPriority(): self
+    {
+        return $this->withPriority(self::PRIORITY_CONSERVE_POWER);
+    }
+
+    public function withPriority(string $priority): self
+    {
+        $config = clone $this;
+        $config->config['headers'] = $config->config['headers'] ?? [];
+        $config->config['headers']['apns-priority'] = $priority;
 
         return $config;
     }